fixed overflow with the extra ft_atoi_with_overflow function, also expanded the unit tester a little

+++ libft/convert/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: lvan-bus <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/03 14:42:23 by lvan-bus      #+#    #+#                 */
-/*   Updated: 2022/12/08 14:53:58 by lvan-bus      ########   odam.nl         */
+/*   Updated: 2023/01/20 15:28:00 by lvan-bus      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
-	int			i;
-	int			sign;
-	long long	result;
+	int	i;
+	int	sign;
+	int	result;
 
 	sign = 1;
 	i = 0;
@@ -26,7 +26,7 @@ long long	ft_atoi(const char *str)
 		if (str[i] == '-')
 			sign *= -1;
 		i++;
-	}	
+	}
 	while (str[i] >= '0' && str[i] <= '9')
 	{
 		result = result * 10 + str[i] - '0';
